move: reworked move function
move() calls findpiece() which creates a list of all the pieces
allowed to do the given move. if there is at the end other than
one piece it results in an error. The function hasvalidmove() has
to be done and must return true if a piece can go from From to To.